feat(eval): support float quiz scores in progress
Change SectionStatus.quiz_score from Optional[int] to Optional[float]:
- Allows recording partial credit scores from AI evaluation
- Supports scores like 6.5/8 instead of only whole numbers
- Compatible with existing int scores through Python's type system

+++ tools/edps/progress.py
@@ -26,7 +26,7 @@ class SectionStatus:
     """Result of checking section completion."""
     is_complete: bool
     recall_score: Optional[int]
-    quiz_score: Optional[int]
+    quiz_score: Optional[float]
 
 
 def check_recall_completion(recall_path: Path) -> RecallResult:
